+++ model/schema_flag_test.go
@@ -1,0 +1,14 @@
+package model
+
+import "testing"
+
+func TestFlag(t *testing.T) {
+	// TODO: Schema.SetFlag
+	// TODO: Schema.SetFlagAt
+	// TODO: Schema.SetTxFlag
+	// TODO: Schema.SetTxFlagAt
+	// TODO: Schema.SetFlags
+	// TODO: Schema.SetFlagsAt
+	// TODO: Schema.SetTxFlags
+	// TODO: Schema.SetTxFlagsAt
+}