+++ model/schema_toggle_test.go
@@ -1,0 +1,14 @@
+package model
+
+import "testing"
+
+func TestToggle(t *testing.T) {
+	// TODO: Schema.ToggleID
+	// TODO: Schema.ToggleIDAt
+	// TODO: Schema.ToggleCode
+	// TODO: Schema.ToggleCodeAt
+	// TODO: Schema.ToggleTxID
+	// TODO: Schema.ToggleTxIDAt
+	// TODO: Schema.ToggleTxCode
+	// TODO: Schema.ToggleTxCodeAt
+}